+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import LoginPage from "./pages/LoginPage";
 import SignupPage from "./pages/SignupPage";
-import RegisterPage from "./components/pages/RegisterPage";
+import RegisterPage from "./pages/RegisterPage";
 
 const App = () => {
   return (
