--- src/components/pages/RegisterPage.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import RegisterForm from "../organisms/RegisterForm"
-
-const RegisterPage = () => {
-    return (
-        <div className="w-screen h-screen flex items-center justify-center ">
-            <RegisterForm/>
-        </div>
-    );
-};
-
-export default RegisterPage;
