+++ src/pages/RegisterPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import RegisterForm from "../organisms/RegisterForm"
+import RegisterForm from '../components/organisms/RegisterForm';
 
 const RegisterPage = () => {
     return (
